% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_kols.R
\name{pick_kols}
\alias{pick_kols}
\title{Pick key opinion leaders from a network given constraints}
\usage{
pick_kols(
  network,
  tosource = TRUE,
  goal = "diffusion",
  m = 1,
  min = 1,
  max = 1,
  top = NULL,
  include = NULL,
  exclude = NULL,
  attribute = NULL,
  best = NULL,
  file = NULL
)
}
\arguments{
\item{network}{a unipartite unweighted network as an adjacency \code{matrix} or \code{\link{igraph}} object}

\item{tosource}{logical: edges point \emph{toward} a source of information}

\item{goal}{string: goal for the KOL team (either \code{"diffusion"} or \code{"adoption"})}

\item{m}{integer: KOL team centrality parameter (\code{m == 1} is equivalent to simple degree centrality)}

\item{min}{integer: minimum KOL team size}

\item{max}{integer: maximum KOL team size}

\item{top}{numeric: restrict scope to the \code{top} nodes with the highest degree, closeness, or betweenness (useful for large networks)}

\item{include}{vector: names or indices of nodes that \strong{must} be included on the KOL team}

\item{exclude}{vector: names or indices of nodes that \strong{can not} be included on the KOL team}

\item{attribute}{string or vector: if \code{network} is an \code{igraph} object, the name of a categorical node attribute recorded as a factor. if \code{network} is an adjacency matrix, a factor containing a node attribute.}

\item{best}{numeric: limit the results to the \code{best} best KOL teams, based on their overall evaluation}

\item{file}{string: filename to write list of possible KOL teams as a CSV}
}
\value{
A list containing a data frame of possible KOL teams with their characteristics, the \code{network}, \code{m}, \code{goal}, and (optionally) \code{attribute}
}
\description{
Pick key opinion leaders from a network given constraints
}
\details{
When seeking to diffuse a piece of information or encourage adoption of a behavior, it is often useful
to recruit the assistance of \emph{key opinion leaders} (KOL) in a network. \code{pick_kols} facilitates selecting
members of a KOL team by returning a dataframe of possible teams. The selection of a KOL team often depends on
several factors, which this function summarizes as ABCDE:
\itemize{
\item Availability - The availability of individuals to serve as a KOL. This can be controlled by the \code{include} and \code{exclude} parameters.
\item Breadth - The fraction of non-KOLs that the KOL team can influence. When \code{goal=="diffusion"}, breadth is measured as the fraction of non-KOLs that a KOL team can reach in \code{m} steps (i.e., m-reach). When \code{goal=="adoption"}, breadth is measured as the fraction of non-KOLs that are directly connected to at least \code{m} KOLs (i.e., m-contact).
\item Cost - The number of KOLs to be recruited and trained (i.e., team size).
\item Diversity - The diversity of members of a KOL team with respect to a categorical \code{attribute} using Simpson's Index.
\item Evaluation - Among teams composed of available individuals, there are many ways to balance considerations of breadth,
cost, and diversity. \code{pick_kols}'s evaluation and ranking of teams uses (breadth x diversity)/cost, which is then
rescaled to a 0-1 range.
}

If \code{network} is a directed network, then \code{tosource} controls how the direction of edges is interpreted:
\itemize{
\item \code{tosource = TRUE} (default) - An edge i -> j is interpreted as "i gets information from j" or "i is influenced
by j" (i.e., the edge points \emph{toward} a source of information or influence). This type of data usually results from
asking respondents to nominate the people from whom they seek advice. In this case, actors with high \emph{in-degree} like
\emph{j} are generally better KOLs.
\item \code{tosource = FALSE} - An edge i -> j is interpreted as "i sends information to j" or "i influences j" (i.e., the
edge points \emph{away} from a source of information or influence). This type of data usually results from asking respondents
to report the people to whom they give advice. In this case, actors with high \emph{out-degree} like \emph{i} are generally better KOLs.
}
}
\examples{
network <- igraph::sample_smallworld(1,26,2,.2)  #An example network
igraph::V(network)$name <- letters[1:26]  #Give the nodes names
igraph::V(network)$gender <- factor(sample(c("M","F"),26,replace=TRUE))  #Give the nodes a "gender"
teams <- pick_kols(network,              #Find KOL teams in `network`
                   m = 2,
                   min = 2,              #containing 2-4 members
                   max = 4,
                   attribute = "gender", #that are gender diverse
                   goal = "diffusion")   #and can help diffuse information
teams$teams[1:10,]  #Look at the top 10 teams
}
